auth-service/requirements.txt
flask
requests
flask_sqlalchemy
psycopg2-binary
python-dotenv
flask-migrate
